Move mixer tests to their own file
This is just because that makes it easier for me to navigate around. No
other reason.
